requirements.txt
django 
djangorestframework 
pydub 
transformers 
sentence-transformers